# api.yaml
swagger: '2.0'
info:
  description: REST API description of the ElasTest TORM Module.
  version: 1.0.0
  title: ElasTest Torm REST API
  contact: {}
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: 'localhost:8091'
basePath: /v2
...
paths:
...
  /tjob:
    get:
      tags:
        - tjob
      summary: Returns all tjobs.
      description: Returns all tjobs for a user loged.
      operationId: getAllTJobsUsingGET
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          description: Successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/TJob'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Resource not found
    post:
      tags:
        - tjob
      summary: Create a new tjob.
      description: Creates a new tjob with the received information.
      operationId: createTJobUsingPOST
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Tjob object that needs to create
          required: true
          schema:
            $ref: '#/definitions/TJob'
      responses:
        '200':
          description: TJob Creation OK
          schema:
            $ref: '#/definitions/TJob'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: Invalid input
    put:
      tags:
        - tjob
      summary: Modifies a existing tjob.
      description: Modifies a existing tjob received as a parameter.
      operationId: modifyTJobUsingPUT
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Tjob object that needs to modify.
          required: true
          schema:
            $ref: '#/definitions/TJob'
      responses:
        '200':
          description: TJob Mofification OK
          schema:
            $ref: '#/definitions/TJob'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: Invalid input
          schema:
            $ref: '#/definitions/TJob'
  '/tjob/{tJobId}':
    get:
      tags:
        - tjob
      summary: Returns a TJob.
      description: Returns the TJob for a fiven TJobId. Returns its detail information.
      operationId: getTJobByIdUsingGET
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: tJobId
          in: path
          description: ID of tJob to retrieve.
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: Successful operation
          schema:
            $ref: '#/definitions/TJob'
        '400':
          description: TJob not found.
          schema:
            $ref: '#/definitions/TJob'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
        - tjob
      summary: Deletes a TJob.
      description: Delete the TJob that matches with a given a tJobId.
      operationId: deleteTJobUsingDELETE
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: tJobId
          in: path
          description: ID of tJob to delete.
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: Deleteted successful
          schema:
            type: integer
            format: int64
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: TJob not found
          schema:
            type: integer
            format: int64
  '/tjob/{tJobId}/exec':
    get:
      tags:
        - tjob execution
      summary: Returns all TJob Executions of a Tjob.
      description: Returns all TJob Executions of a Tjob.
      operationId: getTJobsExecutionsByTJobUsingGET
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: tJobId
          in: path
          description: TJob Id.
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: Successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/TJobExecution'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: TJobs Executions not found
    post:
      tags:
        - tjob execution
      summary: Executes a TJob.
      description: Runs the TJob for a give tJobId and returns a Tjob Execution Id.
      operationId: execTJobUsingPOST
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: tJobId
          in: path
          description: TJob Id.
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: Successful operation
          schema:
            $ref: '#/definitions/TJobExecution'
        '201':
          description: Created
        '400':
          description: Invalid ID supplied
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: TJob not found
        '500':
          description: Server Error
  '/tjob/{tJobId}/exec/{tJobExecId}':
    get:
      tags:
        - tjob execution
      summary: Returns a TJob Execution.
      description: Returns the TJob Execution for a given id.
      operationId: getTJobsExecutionUsingGET
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: tJobId
          in: path
          description: TJob Id.
          required: true
          type: integer
          format: int64
        - name: tJobExecId
          in: path
          description: TJob Execution Id.
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: Successful operation
          schema:
            $ref: '#/definitions/TJobExecution'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: TJob Execution not found
    delete:
      tags:
        - tjob execution
      summary: Deletes a TJob Execution.
      description: Deletes the TJob Execution for a given id. Returns the id.
      operationId: deleteTJobExecutionUsingDELETE
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: tJobId
          in: path
          description: TJob Id.
          required: true
          type: integer
          format: int64
        - name: tJobExecId
          in: path
          description: TJob Execution Id.
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: Successful operation
          schema:
            type: integer
            format: int64
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: TJob Execution not found
          schema:
            type: integer
            format: int64
...
definitions:
  Project:
    type: object
    required:
      - id
      - name
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      suts:
        type: array
        items:
          $ref: '#/definitions/SutSpecification'
      tJobs:
        type: array
        items:
          $ref: '#/definitions/TJob'
      tOJobs:
        type: array
        items:
          $ref: '#/definitions/TOJob'
      tjobs:
        type: array
        items:
          $ref: '#/definitions/TJob'
      tojob:
        type: array
        items:
          $ref: '#/definitions/TOJob'
      tojobs:
        type: array
        items:
          $ref: '#/definitions/TOJob'
  Log:
    type: object
    required:
      - logType
      - logUrl
    properties:
      id:
        type: integer
        format: int64
      logType:
        type: string
        enum:
          - SutLog
          - TestLog
      logUrl:
        type: string
      tJobExec:
        $ref: '#/definitions/TJobExecution'
  SutExecution:
    type: object
    properties:
      deployStatus:
        type: string
        enum:
          - deploying
          - deployed
          - undeploying
          - undeployed
          - error
      id:
        type: integer
        format: int64
        example: 12345678
      sutSpecification:
        $ref: '#/definitions/SutSpecification'
  TJobExecution:
    type: object
    required:
      - duration
      - id
      - result
    properties:
      duration:
        type: integer
        format: int64
      error:
        type: string
      id:
        type: integer
        format: int64
      logs:
        type: array
        description: URLs of logs
        items:
          $ref: '#/definitions/Log'
      result:
        type: string
        enum:
          - SUCCESS
          - FAILURE
          - IN PROGRESS
      sutExecution:
        $ref: '#/definitions/SutExecution'
      tjob:
        $ref: '#/definitions/TJob'
      tojobExecution:
        $ref: '#/definitions/TOJobExecution'
      tojobExecutionb:
        $ref: '#/definitions/TOJobExecution'
  TOJobExecution:
    type: object
    required:
      - id
      - status
      - tojob
    properties:
      id:
        type: integer
        format: int64
      status:
        type: string
        enum:
          - SUCCESS
          - FAILURE
          - IN PROGRESS
      tJobsExecutions:
        type: array
        items:
          $ref: '#/definitions/TJobExecution'
      tOJob:
        $ref: '#/definitions/TOJob'
      tjobsExecutions:
        type: array
        items:
          $ref: '#/definitions/TJobExecution'
      tojob:
        $ref: '#/definitions/TOJob'
  TOJob:
    type: object
    required:
      - name
      - project
      - toscript
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      project:
        $ref: '#/definitions/Project'
      tOScript:
        type: string
      tojobExecutions:
        type: array
        items:
          $ref: '#/definitions/TOJobExecution'
      toscript:
        type: string
        example: 'tjob1 {sut1, browser}, tjob2 {sut1, browser}, tjob3 {sut2, browser2}'
...
  DeployedSut:
    type: object
    required:
      - url
    properties:
      sshKey:
        type: string
      sutSpecification:
        $ref: '#/definitions/SutSpecification'
      url:
        type: string
  SutSpecification:
    type: object
    required:
      - name
      - project
      - specification
    properties:
      deployedSut:
        type: array
        items:
          $ref: '#/definitions/DeployedSut'
      description:
        type: string
        example: This is a SuT description example
      id:
        type: integer
        format: int64
      name:
        type: string
        example: sut definition 1
      project:
        example: '12345678'
        $ref: '#/definitions/Project'
      specification:
        type: string
      sutExecution:
        type: array
        items:
          $ref: '#/definitions/SutExecution'
      tJobs:
        type: array
        items:
          $ref: '#/definitions/TJob'
      tjobs:
        type: array
        items:
          $ref: '#/definitions/TJob'
...
  TJob:
    type: object
    required:
      - name
    properties:
      id:
        type: integer
        format: int64
      imageName:
        type: string
        example: edujgurjc/torm-test-01
      name:
        type: string
        example: testApp1
      project:
        example: '{ id:"1" }'
        $ref: '#/definitions/Project'
      sut:
        example: '{ id:"1" }'
        $ref: '#/definitions/SutSpecification'
      tjobExecs:
        type: array
        items:
          $ref: '#/definitions/TJobExecution'